Primitive highlighting functionality.
Initial implementation of the highlighting functionality. Currently
based around the styling added by the Google source code pretty print
library.

This implementation is not perfect, it runs into issues when multiple
HTML tags are lined up in a row, i.e.:

```
<br /><span><a href="foo.html>
```

In these situations, the library places all of the tags into the `<span
class="tag">...</span>` element, which is what is being highlighted.

The current implementation highlights none of the items. Another
possibility would be to hightlight all of the items if one of the tags
were the desired tag. Also not an ideal situation.

Considering using the Nokogiri representation to build a tree structure
of the HTML source and apply some styling there, but that may be
extremely difficult to do while maintaining the original source (have to
use xmp tag to prevent the HTML source from rendering).