Use more liberal highlighting scheme.
The previous highlighting scheme only worked if the tag was by itself on
a line, or did not immediately touch other tags. This is because I used
the classes provided by the prettyprint library to help break up the
source code. I also think it is more reasonable to highlight an entire
chunk of tags if one is desired than to not highlight anything if
desired tag is not all by its lonesome.

Previously, selecting `span` for highlight would _not_ have highlighted
this block: `<div><span>`. Now, it _will_ highlight the _entire_ block
instead of nothing at all. I think this is more useful for someone who
is scanning the source code.